update
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@1101 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/asr_tts/mod_cepstral/mod_cepstral.c
@@ -179,7 +179,7 @@ static switch_status cepstral_speech_feed_tts(switch_speech_handle *sh, char *te
 static switch_status cepstral_speech_read_tts(switch_speech_handle *sh,
 											  void *data,
 											  size_t *datalen,
-											  size_t *rate,
+											  uint32_t *rate,
 											  switch_speech_flag *flags) 
 {
 	cepstral_t *cepstral;
